+++ .github/workflows/prepare-backport-patch.yml
@@ -5,6 +5,7 @@ on:
       package:
         type: choice
         options:
+        - opentelemetry-instrumentation-genai-agno
         - opentelemetry-instrumentation-genai-anthropic
         - opentelemetry-instrumentation-google-genai
         - opentelemetry-instrumentation-genai-langchain
